chore: move function to mrf utils
